Change extension so that it *only* embeds a (remote) iframe
This comes because `gapi` (the Google APIs) really only work in a real web
page. They look for `postMessage` and `.parent` and hardcoded protocols like
`chrome-extension` and `https` (not `vscode-webview`).

So getting `gapi` to work in the webview context will require some kind of
iframe shim around it, with a lot more work.

In the meantime, we have good enough embedding support to just stick pyret-vmt
as an embedded instance (and eventuall the embed page for code.pyret.org).