Make cookies be samesite: strict
This makes it so session cookies won't be enabled when iframed.

This avoids thinking about a whole host of things that can happen for embedded
instances trying to save files, etc. Embedded instances will *always* behave as
if logged out to start, and will not show any UI elements for logging in.