Starting to fill in the fs library with rpcs
If we have a sendRpc target (like VScode) we can try to fill in the fs library
to some extent.

We won't be able to get full coverage because we can't do synchronous RPCs (so
e.g. filelib won't work transparently)

But we can do a lot of it, and we can probably update the pyret-lang files to
use a reasonable subset that we can polyfill with RPCs to embedding contexts
that do have a filesystem