Add hideInteractions as well as hideDefinitions
hideInteractions is special: we don't have a good way to show run results
(other than highlighting and gutter marks).

So, as soon as run is clicked, it shows interactions again. But this is useful
for having e.g. many editors open in different extension tabs, or just
embedding a plain editor.