Add publicOnly as a flag when initializing Drive storage
This is the start of some refactoring and cleanup around page loads.

When a main CPO window starts up, it unconditionally makes a few requests to
figure out the login state and where the user's programs should be saved,
cached, etc within Drive.

In embedded or VScode contexts, this makes no sense at all: the cookie policy,
API settings are completely turned off any private access to drive. (Like, at a
browser level we leave on the settings that prevent the iframe from setting
cookies)

As a result, there can be really confusing error messages in the console where
stuff just doesn't load and there are a bunch of apparent 404's. Not great when
you're polluting the console of the embedded page for debugging reasons,
either.

So this is the start of some work to more programmatically “turn off” the
stuff that wouldn't work by design anyway to have things look a little
cleaner/do less useless work.