Basic url-file implementation for CPO
This expects the `fs.exists` message to be present on the embedder to check for
file presence.

The option is templated in for now, which the extension has control over when
it first renders the page. Pages may need to be reloaded for the settings
change to take effect.

I noticed/remembered while doing this that the CPO file import does not track
working directories, this should be fixed by adding to the module finder
context just like in cli-module-loader.arr in pyret-lang.