crossorigin=anonymous on PYRET loads
VScode-in-browser has strict content security policy rules.

One of them requires that we serve /editor with certain cross-origin rules (see
server.js)

I didn't fully tie the knot on making sure that if PYRET is cross-origin (which
it is for pyret-horizon!) the requests from preload and dynamic loads check the
right header boxes.