Quality of life improvement for JS errors that leak
This would give the “unexpected exception check console” message, but it
wouldn't always get to the console because of errors in the rendering of the JS
error!

This makes sure the JS error (a) shows up in the error message, which is
ultimately more helpful and (b) is logged in full detail with its stack trace
attached.