Add a param for #theme
This allows embedders to control the theme. The hash param theme overrides the
user setting, but does not change the stored user setting.

The related change in CPO controls this – instead of eagerly changing the user
setting to the set param value, instead let them differ until the user
explicitly changes it with a UI action.